ai_backend_scaffold_pro/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
celery==5.4.0
pillow==10.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
requests==2.32.3
pytest==8.2.1
httpx==0.27.0
